refactor: delegating responsabilities to another componentes
